Fold the widget-styling helpers into the theme header
StatusStyle.h existed for one function, and ButtonStyle.h next to it for another,
both of which apply the theme to a widget instead of returning a string to apply.
They belong with the styles they use, so both move into ThemeManager.h - the same
consolidation UIConstants.h got - and the two headers are gone. Call sites keep
their spelling: the functions stay in geck::ui.
